Merges

+++ users/views.py
@@ -18,7 +18,8 @@ class SignUpView(APIView):
 
 
 class UserDetailView(APIView):
-    # permission_classes = [permissions.IsAuthenticated]
+    permission_classes = [permissions.IsAuthenticated]
+    
     def get(self, request, user_id):
         user = get_object_or_404(User, id=user_id)
         serializer = UserSerializer(user)
@@ -26,14 +27,29 @@ class UserDetailView(APIView):
 
     def put(self, request, user_id):
         user = get_object_or_404(User, id=user_id)
-        serializer = UserDetailSerializer(user, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
+        # print(str(request.user), str(user.email))
+        if str(request.user) == user.email:
+            serializer = UserDetailSerializer(user, data=request.data)
+            if serializer.is_valid():
+                serializer.save()
+                return Response(serializer.data, status=status.HTTP_200_OK)
+            else:
+                return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
+            return Response("권한이 없습니다", status=status.HTTP_403_FORBIDDEN)
+        # serializer = UserDetailSerializer(user, data=request.data)
+        # if serializer.is_valid():
+        #     serializer.save()
+        #     return Response(serializer.data, status=status.HTTP_200_OK)
+        # else:
+        #     return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
     def delete(self, request, user_id):
         user = get_object_or_404(User, id=user_id)
-        user.delete()
-        return Response(status=status.HTTP_204_NO_CONTENT)
+        if str(request.user) == user.email:
+            user.delete()
+            return Response(status=status.HTTP_204_NO_CONTENT)
+        else:
+            return Response("권한이 없습니다", status=status.HTTP_403_FORBIDDEN)
+        # user.delete()
+        # return Response(status=status.HTTP_204_NO_CONTENT)
